Remove Home page from rendering
